requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart